Drop requirement that output mask cannot contain duplicate entries
This allows output indices to depend on the same input axis.
This allows for example the creation of views on diagonal or subdiagonal
entries of a matrix:

```lisp
(compute (lazy-reshape (lazy-array #2A((2 1 0)
                                       (2 3 0)
                                       (2 1 5)))
                       (~ 3)
                       (τ (i) (i i))))
-> #(2 3 5)

(compute (lazy-reshape (lazy-array #2A((2 1 0)
                                       (2 3 0)
                                       (2 1 5)))
                       (~ 2)
                       (τ (i) ((1+ i) i))))
-> #(2 1)
```

Such transforms are not invertible of course, but Petalisp detects that
if you try to invert them anyways.

Use of the transforms in reshape will repeat the respective dimension in
another axis
```
(compute (reshape #(2 1 0) (τ (i) (i i))))
-> #2A((2 1 0) (2 1 0) (2 1 0))

(compute (reshape #2A((3 4 5)
                     (2 1 0)) (τ (i j) (i j i))))
-> #3A(((3 2) (4 1) (5 0)) ((3 2) (4 1) (5 0)))
```